src/requirements.txt
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
greenlet==3.1.1
idna==3.10
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pytz==2025.1
PyYAML==6.0.1
requests==2.32.3
SQLAlchemy==2.0.37
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
